Implement Coverflow without using zIndex
The zIndex implemention of Android doesn't work properly when the
children list is changed dynamicllay. The Coverflow worked well in
both iOS and Android with zIndex, but started giving problem in
android when the children were removed/added dynamically on the list.

To get around this issue, the children were reordered based on
current selection and rerendered.